requirements.txt
anyio==4.3.0
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instaloader==4.11
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-telegram-bot==21.1.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
